Move status center logic to view model
