+++ Files/Interacts/IStatusCenterActions.cs
@@ -1,5 +1,6 @@
 ﻿using Files.Enums;
 using Files.UserControls;
+using Files.ViewModels;
 using System;
 
 namespace Files.Interacts
