+++ Files/UserControls/StatusBarControl.xaml
@@ -14,7 +14,8 @@
     xmlns:xh="using:Files.Helpers.XamlHelpers"
     d:DesignHeight="40"
     d:DesignWidth="400"
-    mc:Ignorable="d">
+    mc:Ignorable="d"
+    Loading="UserControl_Loading">
     <UserControl.Resources>
         <ResourceDictionary>
             <converters1:BoolNegationConverter x:Key="BoolNegationConverter" />
@@ -112,7 +113,8 @@
                                 MinHeight="300"
                                 MaxHeight="400"
                                 x:FieldModifier="public"
-                                IsTabStop="True" />
+                                IsTabStop="True" 
+                                StatusCenterViewModel="{x:Bind StatusCenterViewModel}"/>
                         </Flyout>
                     </Button.Flyout>
                 </Button>
@@ -120,10 +122,10 @@
         </StackPanel>
 
         <Interactivity:Interaction.Behaviors>
-            <Core:DataTriggerBehavior Binding="{x:Bind OngoingTasksControl.AnyOperationsOngoing, Mode=OneWay}" Value="False">
+            <Core:DataTriggerBehavior Binding="{x:Bind StatusCenterViewModel.AnyOperationsOngoing, Mode=OneWay}" Value="False">
                 <Core:GoToStateAction StateName="NoOperations" />
             </Core:DataTriggerBehavior>
-            <Core:DataTriggerBehavior Binding="{x:Bind OngoingTasksControl.AnyOperationsOngoing, Mode=OneWay}" Value="True">
+            <Core:DataTriggerBehavior Binding="{x:Bind StatusCenterViewModel.AnyOperationsOngoing, Mode=OneWay}" Value="True">
                 <Core:GoToStateAction StateName="InProgress" />
             </Core:DataTriggerBehavior>
         </Interactivity:Interaction.Behaviors>
