+++ Files/UserControls/StatusCenter.xaml
@@ -4,6 +4,7 @@
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:local="using:Files.UserControls"
+    xmlns:models="using:Files.ViewModels"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:muxc="using:Microsoft.UI.Xaml.Controls"
     d:DesignHeight="300"
@@ -130,10 +131,10 @@
             Grid.Row="1"
             VerticalAlignment="Stretch"
             ItemContainerStyle="{StaticResource ListViewItemStyleFixed}"
-            ItemsSource="{x:Bind local:StatusCenter.StatusBannersSource, Mode=OneWay}"
+            ItemsSource="{x:Bind models:StatusCenterViewModel.StatusBannersSource, Mode=OneWay}"
             SelectionMode="None">
             <ListView.ItemTemplate>
-                <DataTemplate x:DataType="local:StatusBanner">
+                <DataTemplate x:DataType="models:StatusBanner">
                     <Grid
                         Padding="4"
                         HorizontalAlignment="Stretch"
